refactor : User와 연결 및 후원 요청 검증 보완(#15)
