+++ src/main/java/com/pawever/server/domain/donation/repository/DonationRepository.java
@@ -4,7 +4,7 @@ import com.pawever.server.domain.donation.entity.Donation;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-import com.pawever.server.domain.user.entity.User;
+import com.pawever.server.domain.user.entity.jpa.User;
 
 import java.util.List;
 import java.util.Optional;
